James hit a brick wall at 3:22pm

+++ tenmo-server/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -10,7 +10,7 @@ public class Account {
 
     public Account() {}
 
-    public Account(long userId, BigDecimal balance) {
+    public Account(int userId, BigDecimal balance) {
         this.userId = userId;
         this.balance = balance;
     }
